Fix missing backingField mapping and redundant Int64 cast
- Add backingField case to symbolKindFromMetadataToken so backing field
  symbols deserialize correctly from metadata instead of returning nil
- Remove redundant Int64() cast in CallLowerer+SafeMemberCalls; defaultMask
  is already Int64, matching the direct usage in CallLowerer+MemberCalls

+++ Sources/CompilerCore/KIR/CallLowerer+SafeMemberCalls.swift
@@ -113,7 +113,7 @@ extension CallLowerer {
                 arguments: &finalArguments
             )
             appendDefaultMaskArgument(
-                Int64(safeNormalized.defaultMask),
+                safeNormalized.defaultMask,
                 sema: sema,
                 arena: arena,
                 instructions: &instructions.instructions,
